+++ app/ui/src/pages/CurrencyPage.jsx
@@ -1,0 +1,9 @@
+import Currency from "../components/Currency/Currency";
+
+export const CurrencyPage = () => {
+  return (
+    <Currency />
+  );
+}
+
+export default CurrencyPage;